@@ -38,11 +38,11 @@
     "lodash-es": "^4.17.21",
     "npm-run-all": "^4.1.5",
     "qs": "^6.10.3",
-    "react": "^17.0.1",
+    "react": "^18.3.1",
     "react-click-outside-hook": "^1.1.1",
     "react-confetti": "^6.1.0",
     "react-datepicker": "^4.1.1",
-    "react-dom": "^17.0.1",
+    "react-dom": "^18.3.1",
     "react-icons": "^4.2.0",
     "react-idle-timer": "^5.7.2",
     "react-loader-spinner": "^4.0.0",
